test: fix jwt claims type assertions

+++ tests/auth_register_login_test.go
@@ -51,9 +51,9 @@ func TestRegisterLogin_LoginHappyPath(t *testing.T) {
 	claims, ok := tokenParsed.Claims.(jwt.MapClaims)
 	require.True(t, ok)
 
-	assert.Equal(t, regResp.UserId, int64(claims["uid"].(float64)))
+	assert.Equal(t, regResp.UserId, claims["uid"].(int64))
 	assert.Equal(t, email, claims["email"].(string))
-	assert.Equal(t, appID, claims["app_id"].(float64))
+	assert.Equal(t, appID, claims["app_id"].(int))
 
 	const deltaSeconds = 1
